Cleans up helpers changes
